% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{call_main}
\alias{call_main}
\title{Call `main()`}
\usage{
call_main()
}
\description{
Call the function `main()` with the script name and the command line
argument. When the `main()` finishes, exit with the return code as 
the program status.
}
\examples{
\dontrun{
    if (sys.nframe() == 0) {
        call_main()
    }
}

}
