% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{clip_vals}
\alias{clip_vals}
\title{Clips numeric values so that they lie in the range [min_val, max_val]
Works on both numeric arrays and dataframes}
\usage{
clip_vals(x, min_val = NULL, max_val = NULL)
}
\description{
Clips numeric values so that they lie in the range [min_val, max_val]
Works on both numeric arrays and dataframes
}
