% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{colMins}
\alias{colMins}
\title{A wrapper for matrixStats::colMins()
Returns the minimal numeric value per column in a matrix or a dataframe}
\usage{
colMins(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[matrixStats:rowMaxs]{matrixStats::rowMaxs}}
  \describe{
    \item{\code{x}}{An NxK \code{\link[base]{matrix}} or, if \code{dim.} is specified,
an N * K \code{\link[base]{vector}}.}
    \item{\code{rows}}{A \code{\link[base]{vector}} indicating subset of rows to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}
    \item{\code{cols}}{A \code{\link[base]{vector}} indicating subset of columns to
operate over. If \code{\link[base]{NULL}}, no subsetting is done.}
    \item{\code{na.rm}}{If \code{\link[base:logical]{TRUE}}, missing values are
excluded.}
    \item{\code{dim.}}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.  \emph{Comment:} The reason for this argument
being named with a period at the end is purely technical (we get a run-time
error if we try to name it \code{dim}).}
    \item{\code{useNames}}{If \code{\link[base:logical]{TRUE}} (default), names
attributes of the result are set, otherwise not.}
  }}
}
\description{
A wrapper for matrixStats::colMins()
Returns the minimal numeric value per column in a matrix or a dataframe
}
