% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutree_order}
\alias{cutree_order}
\title{Cut a tree into groups of data in the order of the tree}
\usage{
cutree_order(tree, k = k, h = h)
}
\arguments{
\item{tree}{an hclust object}

\item{k}{the desired number of groups}

\item{h}{height where the tree is to be cut}
}
\description{
Cut a tree (result from hclust) into groups of data. Groups are in the order of the tree leafs
}
\section{Notes}{
 Taken from http://larmarange.github.io/JLutils/reference/cutree.order.html
}

\examples{

hc <- hclust(dist(USArrests))
memb <- cutree_order(hc, k = 10)
}
