% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exit}
\alias{exit}
\title{Equivalent to python's sys.exit()
Stop the script without saving R's status.}
\usage{
exit(...)
}
\arguments{
\item{rc}{Exist status. A NULL is equivalent to 0. Non integer values will be printed to stderr and exit status will be set to 1.}
}
\description{
Equivalent to python's sys.exit()
Stop the script without saving R's status.
}
