% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_csv}
\alias{get_csv}
\title{Wrap a function that results a data frame to cache results in a csv file}
\usage{
get_csv(func, fn, row.names = FALSE, ...)
}
\arguments{
\item{func}{function to wrap}

\item{fn}{name of csv file}

\item{row.names}{Should row names be cached as well}

\item{...}{default parameters for func}
}
\value{
a wrapped function
}
\description{
Wrap a function that results a data frame to cache results in a csv file
}
\examples{

 temp_file <- tempfile(fileext = ".csv")

 calc_mtcars_cyl <- function(cyl) {
      message("calculating")
      mtcars[mtcars$cyl == cyl, ]
 }

get_mtcars_cyl <- get_csv(calc_mtcars_cyl, temp_file)

# Call "calc_mtcars_cyl" function
get_mtcars_cyl(6)

# Load cached file
get_mtcars_cyl(6)

# Force re-calculating
get_mtcars_cyl(6, recalc=TRUE)

}
