% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rds}
\alias{get_rds}
\title{Wrap a function to cache results in an rds file}
\usage{
get_rds(func, fn, ...)
}
\arguments{
\item{func}{function to wrap}

\item{fn}{name of rds file}

\item{...}{default parameters for func}
}
\value{
a wrapped function
}
\description{
Wrap a function to cache results in an rds file
}
\examples{

 temp_file <- tempfile(fileext = ".rds")

 calc_mtcars_cyl <- function(cyl) {#'       
      mtcars[mtcars$cyl == cyl, ]
 }

get_mtcars_cyl <- get_rds(calc_mtcars_cyl, temp_file)

# Call "calc_mtcars_cyl" function
get_mtcars_cyl(6)

# Load cached file
get_mtcars_cyl(6)

# Force re-calculating
get_mtcars_cyl(6, recalc=TRUE)

}
