% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gps}
\alias{gps}
\title{Get current plot size for jupyter IRKernel}
\usage{
gps()
}
\value{
a vector with the current width and height of the IRKernel plot
}
\description{
Get current plot size for jupyter IRKernel
}
\examples{
gps()

}
