% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{\%cache_matrix\%}
\alias{\%cache_matrix\%}
\title{Tries reading a cached matrix result from the filename on the right
side. If the file is missing, generate the matrix using the function
call on the left side and write the result to the file.}
\usage{
call \%cache_matrix\% fname
}
\arguments{
\item{call}{function call}

\item{fname}{filename}
}
\description{
Tries reading a cached matrix result from the filename on the right
side. If the file is missing, generate the matrix using the function
call on the left side and write the result to the file.
}
\examples{
temp_file <- tempfile(fileext = ".tsv")
  calc_mtcars_cyl <- function(cyl) {
      message("calculating")
      as.matrix(mtcars[mtcars$cyl == cyl, ])
  }

  # Call "calc_mtcars_cyl" function
  res1 <- calc_mtcars_cyl(6) \%cache_matrix\% temp_file

  # Load cached file
  res2 <- calc_mtcars_cyl(6) \%cache_matrix\% temp_file

}
