% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{\%cache_rds\%}
\alias{\%cache_rds\%}
\title{Tries reading a cached dataframe result from the filename on the right
side. If the file is missing, generate the dataframe using the
function call on the left side and write the result to the file. If "tgutil.cache"
option is FALSE, right side would be always recumputed.}
\usage{
call \%cache_rds\% fname
}
\arguments{
\item{call}{function call}

\item{fname}{filename}
}
\description{
Tries reading a cached dataframe result from the filename on the right
side. If the file is missing, generate the dataframe using the
function call on the left side and write the result to the file. If "tgutil.cache"
option is FALSE, right side would be always recumputed.
}
\examples{
temp_file <- tempfile(fileext = "rds")
  calc_mtcars_cyl <- function(cyl) {
      message("calculating")
      mtcars[mtcars$cyl == cyl, ]
  }

  # Call "calc_mtcars_cyl" function
  res1 <- calc_mtcars_cyl(6) \%cache_rds\% temp_file

  # Load cached file
  res2 <- calc_mtcars_cyl(6) \%cache_rds\% temp_file

}
