% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{\%fcache_rds\%}
\alias{\%fcache_rds\%}
\title{cache_rds that force recreating the cached file}
\usage{
call \%fcache_rds\% fname
}
\arguments{
\item{call}{function call}

\item{fname}{filename}
}
\description{
cache_rds that force recreating the cached file
}
\examples{
temp_file <- tempfile(fileext = "rds")
calc_mtcars_cyl <- function(cyl) {
    message("calculating")
    mtcars[mtcars$cyl == cyl, ]
}

# Call "calc_mtcars_cyl" function
res1 <- calc_mtcars_cyl(6) \%cache_rds\% temp_file

# Force re-calling the function
res2 <- calc_mtcars_cyl(6) \%fcache_rds\% temp_file
}
