% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_character}
\alias{load_character}
\title{Read a character vector from a file created by save_character}
\usage{
load_character(fname)
}
\description{
Read a character vector from a file created by save_character
}
