% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_numeric}
\alias{load_numeric}
\title{Read a numeric vector from a file created by save_numeric}
\usage{
load_numeric(fname)
}
\arguments{
\item{fname}{name of file to load from}
}
\description{
Read a numeric vector from a file created by save_numeric
}
