% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{num_physical_cores}
\alias{num_physical_cores}
\title{Return the number of the system's physical CPU cores}
\usage{
num_physical_cores()
}
\description{
Return the number of the system's physical CPU cores
}
