% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pmean}
\alias{pmean}
\title{Vectorized mean, similiar to \code{pmin} and \code{pmax}}
\usage{
pmean(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric vectors to average}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\value{
a vector with mean of \code{...} arguments
}
\description{
Vectorized mean, similiar to \code{pmin} and \code{pmax}
}
