% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_fastq}
\alias{read_fastq}
\title{Read a fastq file to data frame}
\usage{
read_fastq(fn, ...)
}
\arguments{
\item{fn}{of the file - can be zipped (with "gz" file extension).}

\item{...}{other parameters for \code{fread}. Note that \code{cmd,col.names} and \code{sep} are set by the function.}
}
\value{
data frame with "id", "seq", and "qual" corresponding to lines 1-3 of the fastq file
}
\description{
Read a fastq file to data frame
}
