% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{run_wide}
\alias{run_wide}
\title{Similar to mclapply, however ensures a reproducible random-number
stream for each application of fun()}
\usage{
run_wide(x, fun, ..., mc_preschedule = TRUE,
  mc_cores = num_physical_cores(), .seeds = NULL)
}
\description{
Similar to mclapply, however ensures a reproducible random-number
stream for each application of fun()
}
