% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{save_character}
\alias{save_character}
\title{Write a character vector to a file, preserving cell names.}
\usage{
save_character(x, fname)
}
\description{
Write a character vector to a file, preserving cell names.
}
