% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{segment_by}
\alias{segment_by}
\title{Segment a dataframe according to consecutive identical values of a
column expression.}
\usage{
segment_by(df, column, var = "segment")
}
\description{
Segment a dataframe according to consecutive identical values of a
column expression.
}
