% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter-utils.R
\name{set_plot_size}
\alias{set_plot_size}
\title{Set plot size for jupyter IRKernel}
\usage{
set_plot_size(width, height)
}
\arguments{
\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Set plot size for jupyter IRKernel
}
\examples{

# run inside a jupyter notebook
set_plot_size(10, 5)
plot(1:10)

}
