% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter-utils.R
\name{source_jupyter}
\alias{source_jupyter}
\title{Source a jupyter notebook}
\usage{
source_jupyter(file, outdir = NULL, jupytext_path = "jupytext")
}
\arguments{
\item{file}{path to the jupyter notebook}

\item{outdir}{directory to store the converted R file. By default the file would be created in the same directory as the jupyter notebook}

\item{jupytext_path}{path to jupytext executable}
}
\value{
None
}
\description{
convert a jupyter notebook to R using jupytext and source it. Jupytext must be installed and available in the path
}
\examples{
\dontrun{
# notebook.R is created in the same directory as notebook.ipynb
source_jupyter("notebook.ipynb")

# notebook.R is created in a temp directory
source_jupyter("notebook.ipynb", outdir = tempdir())
}

}
