% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter-utils.R
\name{spz}
\alias{spz}
\title{Alias for set_plot_size}
\usage{
spz(width, height)
}
\arguments{
\item{width}{width of the plot}

\item{height}{height of the plot}
}
\description{
Alias for set_plot_size
}
\examples{

# run inside a jupyter notebook
spz(10, 5)
plot(1:10)

}
