% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tgplot_heatmap}
\alias{tgplot_heatmap}
\title{Easily plot a matrix as an heatmap using ggplot2}
\usage{
tgplot_heatmap(mtrx, col_names = NULL, row_names = NULL, xlab = NULL,
  ylab = NULL, plot_top = TRUE, plot_right = TRUE,
  col_names_orient = c("horizontal", "vertical", "slanted"))
}
\arguments{
\item{mtrx}{The matrix to be plotted}

\item{col_names}{The names that will be used as the X-axis labels.
A NULL (the defualt) indicates that the original matrix colnames are used.
A FALSE indicates that no X-axis labels should be displayed.}

\item{row_names}{The names that will be used as the Y-axis labels.
A NULL (the defualt) indicates that the original matrix rownames are used.
A FALSE indicates that no Y-axis labels should be displayed.}

\item{xlab}{The X-axis label}

\item{ylab}{The Y-axis label}

\item{plot_top}{Whether to put the X-axis labels at the top of the heatmap (as well as at the bottom)}

\item{plot_right}{Whether to put the Y-axis labels at the right side of the heatmap (as well as at the left hand side)}

\item{col_names_orient}{Orientation of the X-axis label}
}
\description{
Easily plot a matrix as an heatmap using ggplot2
}
