% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_fastq}
\alias{write_fastq}
\title{Write a data frame with "id", "seq" and "qual" to fastq file}
\usage{
write_fastq(df, fn)
}
\arguments{
\item{df}{data frame with "id", "seq" and "qual" columns (e.g. output of \code{read_fastq})}

\item{name}{of the file - can be zipped (with "gz" file extension).}
}
\value{
None
}
\description{
Write a data frame with "id", "seq" and "qual" to fastq file
}
