% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{colMins}
\alias{colMins}
\title{A wrapper for matrixStats::colMins()
Returns the minimal numeric value per column in a matrix or a dataframe}
\usage{
colMins(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[matrixStats:rowRanges]{matrixStats::rowMaxs}}
  \describe{
    \item{\code{x}}{A \code{\link[base]{numeric}} NxK \code{\link[base]{matrix}}.}
    \item{\code{rows}}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}
    \item{\code{cols}}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}
    \item{\code{na.rm}}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}
    \item{\code{dim.}}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}
  }}
}
\description{
A wrapper for matrixStats::colMins()
Returns the minimal numeric value per column in a matrix or a dataframe
}
