% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ggpreview}
\alias{ggpreview}
\title{Plot using preview instead of plotting to X device}
\usage{
ggpreview(
  plot = ggplot2::last_plot(),
  filename = tempfile(fileext = ".png"),
  ...
)
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{filename}{File name to create on disk.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width}}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
    \item{\code{height}}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
    \item{\code{units}}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
  }}
}
\description{
uses \code{\link[ggplot2]{ggsave}} to save a ggplot object to file and then opens a new device and displays it.
}
\examples{

\dontrun{
 p <- ggplot(mtcars, aes(mpg, wt)) + geom_point() + ggpreview()
}

}
\seealso{
\code{\link[ggplot2]{ggsave}}
}
