% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psum}
\alias{psum}
\title{Vectorized sum, similiar to \code{pmin} and \code{pmax}}
\usage{
psum(..., na.rm = FALSE)
}
\arguments{
\item{...}{numeric vectors to sum}

\item{na.rm}{a logical indicating whether missing values should be removed}
}
\value{
a vector with sum of \code{...} arguments
}
\description{
Vectorized sum, similiar to \code{pmin} and \code{pmax}
}
\examples{
psum(c(1, 2, 3), c(4, 5, 6))
psum(c(1, 2, 3), pi)
psum(c(1, NA, 3), c(4, 5, 6), na.rm = TRUE)

}
