% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{rowMaxs}
\alias{rowMaxs}
\title{A wrapper for matrixStats::rowMaxs()
Returns the maximal numeric value per row in a matrix or a dataframe}
\usage{
rowMaxs(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)
}
\description{
A wrapper for matrixStats::rowMaxs()
Returns the maximal numeric value per row in a matrix or a dataframe
}
