% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{rowMins}
\alias{rowMins}
\title{A wrapper for matrixStats::rowMins()
Returns the minimal numeric value per row in a matrix or a dataframe}
\usage{
rowMins(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)
}
\description{
A wrapper for matrixStats::rowMins()
Returns the minimal numeric value per row in a matrix or a dataframe
}
