% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_x_log2}
\alias{scale_x_log2}
\title{Scale x axis by log2 (similar to scale_x_log10)}
\usage{
scale_x_log2(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_log10}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Scale x axis by log2 (similar to scale_x_log10)
}
