% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_y_log2}
\alias{scale_y_log2}
\title{Scale y axis by log2 (similar to scale_y_log10)}
\usage{
scale_y_log2(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_log10}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Scale y axis by log2 (similar to scale_y_log10)
}
