% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{tgplot_add_axis_annotation}
\alias{tgplot_add_axis_annotation}
\title{Add an axis annotation to a ggplot matrix}
\usage{
tgplot_add_axis_annotation(
  heatmap,
  annotation,
  position = "bottom",
  size = 0.02,
  label = NULL,
  plot_left = TRUE,
  plot_right = TRUE
)
}
\arguments{
\item{heatmap}{a ggplot matrix object - output of \code{tgplot_heatmap}}

\item{annotation}{can be either a vector of colors or a ggplot object}

\item{position}{position of the axis annotation ("top", "bottom", "left" or "right")}

\item{size}{width of the annotation}

\item{label}{annotation label (optional)}

\item{plot_left}{plot a label to the left of the annotation}

\item{plot_right}{plot a label to the right of the annotation}
}
\value{
a gtable object. Can be plotted using \code{cowplot::ggdraw}.
}
\description{
Add an axis annotation to a ggplot matrix
}
\examples{
tgplot_heatmap(as.matrix(mtcars)) \%>\%
    tgplot_add_axis_annotation(as.matrix(mtcars)[5, ] + 1) \%>\%
    cowplot::ggdraw()
}
