% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgutil-package.R
\docType{package}
\name{tgutil-package}
\alias{tgutil}
\alias{tgutil-package}
\title{tgutil: Simple utility functions for Tanay lab code}
\description{
Simple utility functions that are shared across several packages maintained by the Tanay lab.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/tanaylab/tgutil/issues}
}

}
\author{
\strong{Maintainer}: Elad Chomsky \email{elad@chomsky.me}

Authors:
\itemize{
  \item Aviezer Lifshitz \email{aviezer.lifshitz@weizmann.ac.il}
}

}
\keyword{internal}
